Remove CLI code from module entry point
strong-pm, when used as a module, now exports it's Server, and the CLI
code is in the sl-pm.js CLI file.
