+++ test/helper.js
@@ -9,7 +9,6 @@ require('shelljs/global');
 process.env.PATH += path.delimiter
   + path.resolve(__dirname, '../node_modules/.bin');
 
-exports.main = require('../').main;
 exports.prepare = require('../lib/drivers/common/prepare').prepare;
 exports.stop = stop;
